+++ e2e/wdio/core/fixtures/appData/card-content.ts
@@ -1,0 +1,72 @@
+export default {
+    cardTitleArr: [
+        'Card Title',
+        'Card Title',
+        'Card Title',
+        'Card Title'
+    ],
+    cardSubtitle: 'Card Subtitle',
+    badgeText: 'New',
+    cardTypeAttr: 'cardtype',
+    compactAttr: 'ng-reflect-compact',
+    fontSizeAttr: 'font-size',
+    compactFont: '14px',
+    loaderAttr: 'ng-reflect-is-loading',
+    cardListItemText: [
+        'item 1',
+        'item 2',
+        'item 3',
+        'item 4'
+    ],
+    btnText: [
+        'Approve',
+        'Reject'
+    ],
+    analyticsTitle: 'Failure Breakdown',
+    tableHeaderText: [
+        'Name',
+        'Price',
+        'Country',
+        'Status'
+    ],
+    tableContentsText: [
+        'Banana India 5 EUR Available',
+        'Pineapple Mexico 2 EUR Out of stock',
+        'Orange Spain 6 EUR Temporary unavailable'
+    ],
+    tableCardNames: [
+        'Banana',
+        'Pineapple',
+        'Orange'
+    ],
+    tableCardCountries: [
+        'India',
+        'Mexico',
+        'Spain'
+    ],
+    tableCardPrices: [
+        '5 EUR',
+        '2 EUR',
+        '6 EUR'
+    ],
+    tableCardStatuses: [
+        'Available',
+        'Out of stock',
+        'Temporary unavailable'
+    ],
+    statusColors: [
+        '16,126,62',
+        '187,0,0',
+        '5,59,112'
+    ],
+    tableCardTitle: 'Table card',
+    barChartTitle: 'Bar Chart List Card',
+    barChartCounter: '4 of 10',
+    colorAttr: 'color',
+    barChartItems: [
+        'Contract A',
+        'Contract B',
+        'Contract C',
+        'Contract D'
+    ]
+};