+++ e2e/wdio/platform/fixtures/appData/input-group-page-contents.ts
@@ -1,0 +1,18 @@
+export default {
+    standardInputLabels: [
+        'Left aligned text addon',
+        'Right aligned text addon',
+        'Left and right aligned text addons',
+        'Button as an addon',
+        'Icon as an addon'
+    ],
+    standardInputPlaceholders: [
+        'Enter value',
+        'Enter value',
+        'Enter value',
+        'Enter value',
+        'Email'
+    ],
+    inputWithFormLabel: 'Quantity',
+    inputWithFormPlaceholder: 'Things quantity'
+};