+++ e2e/wdio/platform/fixtures/appData/panel-page-content.ts
@@ -1,7 +1,5 @@
-export default
-{
-    expandable_panel_header : 'Panel Title',
-    action_panel_edit_button : 'Edit',
-    action_panel_delete_button : 'Delete',
-
-}
+export default {
+    expandable_panel_header: 'Panel Title',
+    action_panel_edit_button: 'Edit',
+    action_panel_delete_button: 'Delete'
+};