+++ e2e/wdio/platform/fixtures/appData/standard-link-item-contents.ts
@@ -1,0 +1,18 @@
+export default {
+    noBorderAttr: 'noBorder',
+    borderAttr: 'border-bottom-style',
+    compactAttr: 'ng-reflect-compact',
+    byLineAttr: 'ng-reflect-byline',
+    byLineAltAttr: 'ng-reflect-has-by-line',
+    solidStyle: 'solid',
+    noStyle: 'none',
+    linkAttr: 'href',
+    secondaryAttr: 'secondary',
+    secondaryTypes: [
+        'positive',
+        'negative',
+        'neutral',
+        'informative',
+        'critical'
+    ]
+};