+++ e2e/wdio/platform/pages/action-list-item.po.ts
@@ -1,0 +1,19 @@
+import { BaseComponentPo } from './base-component.po';
+import { waitForElDisplayed } from '../../driver/wdio';
+
+export class ActionListItemPo extends BaseComponentPo {
+    url = '/action-list-item';
+    root = '#page-content';
+
+    actionBtns = 'fdp-action-list-item button';
+    actionLists = 'fdp-borderless-action-list-item-example fdp-list';
+    actionSections = 'fdp-borderless-action-list-item-example ul';
+
+    exampleAreaContainersArr = '.fd-doc-component';
+    rtlSwitcherArr = 'rtl-switch .fd-switch__handle';
+
+    open(): void {
+        super.open(this.url);
+        waitForElDisplayed(this.root);
+    }
+}