+++ libs/core/src/lib/pagination/pagination.module.ts
@@ -4,11 +4,12 @@ import { CommonModule } from '@angular/common';
 import { PaginationComponent } from './pagination.component';
 import { ButtonModule } from '../button/button.module';
 import { IconModule } from '../icon/icon.module';
+import { SelectModule } from '../select/select.module';
 import { PaginationService } from './pagination.service';
 
 @NgModule({
     declarations: [PaginationComponent],
-    imports: [CommonModule, ButtonModule, IconModule],
+    imports: [CommonModule, ButtonModule, IconModule, SelectModule],
     providers: [PaginationService],
     exports: [PaginationComponent]
 })