+++ libs/core/src/lib/utils/tests/mobile-mode-view-elements.ts
@@ -14,6 +14,6 @@ export function getMobileModeViewElements(fixture: ComponentFixture<any>): Mobil
         dialogTitle: fixture.nativeElement.querySelector('[fd-title]'),
         dialogFooter: fixture.nativeElement.querySelector('fd-dialog-footer'),
         dialogCloseBtn: fixture.nativeElement.querySelector('[fd-dialog-close-button]'),
-        footerButtons: fixture.nativeElement.querySelectorAll('[fd-dialog-decisive-button]')
+        footerButtons: fixture.nativeElement.querySelectorAll('fd-button-bar')
     }
 }